admin/util: extract out lw_shared_container
This utility class needs to be shared between multiple places.
