+++ src/v/redpanda/CMakeLists.txt
@@ -50,7 +50,7 @@ v_cc_library(
     admin/transaction.cc
     admin/usage.cc
     admin/kafka.cc
-    admin/validation.cc
+    admin/util.cc
     cli_parser.cc
     application.cc
     monitor_unsafe_log_flag.cc
